Refactor countries data file, now only available in components lib
