--- web-components/src/components/inputs/countries.ts
@@ -1,222 +1,0 @@
-export const countries: { [key: string]: string } = {
-  AD: 'Andorra',
-  AE: 'United Arab Emirates',
-  AF: 'Afghanistan',
-  AG: 'Antigua and Barbuda',
-  AI: 'Anguilla',
-  AL: 'Albania',
-  AM: 'Armenia',
-  AN: 'Netherlands Antilles',
-  AO: 'Angola',
-  AR: 'Argentina',
-  AT: 'Austria',
-  AU: 'Australia',
-  AZ: 'Azerbaijan',
-  BA: 'Bosnia and Herzegovina',
-  BB: 'Barbados',
-  BD: 'Bangladesh',
-  BE: 'Belgium',
-  BF: 'Burkina Faso',
-  BG: 'Bulgaria',
-  BH: 'Bahrain',
-  BI: 'Burundi',
-  BJ: 'Benin',
-  BM: 'Bermuda',
-  BN: 'Brunei Darussalam',
-  BO: 'Bolivia',
-  BR: 'Brazil',
-  BS: 'Bahamas',
-  BT: 'Bhutan',
-  BW: 'Botswana',
-  BY: 'Belarus',
-  BZ: 'Belize',
-  CA: 'Canada',
-  CF: 'Central African Republic',
-  CG: 'Congo',
-  CH: 'Switzerland',
-  CI: 'Ivory Coast',
-  CL: 'Chile',
-  CM: 'Cameroon',
-  CN: 'China',
-  CO: 'Colombia',
-  CR: 'Costa Rica',
-  CU: 'Cuba',
-  CV: 'Cape Verde',
-  CY: 'Cyprus',
-  CZ: 'Czech Republic',
-  DE: 'Germany',
-  DJ: 'Djibouti',
-  DK: 'Denmark',
-  DM: 'Dominica',
-  DO: 'Dominican Republic',
-  DZ: 'Algeria',
-  EC: 'Ecuador',
-  EE: 'Estonia',
-  EG: 'Egypt',
-  ER: 'Eritrea',
-  ES: 'Spain',
-  ET: 'Ethiopia',
-  FI: 'Finland',
-  FJ: 'Fiji',
-  FM: 'Micronesia, Federated States of',
-  FO: 'Faroe Islands',
-  FR: 'France',
-  GA: 'Gabon',
-  GB: 'United Kingdom',
-  GD: 'Grenada',
-  GE: 'Georgia',
-  GF: 'French Guiana',
-  GH: 'Ghana',
-  GL: 'Greenland',
-  GM: 'Gambia',
-  GN: 'Guinea',
-  GP: 'Guadeloupe',
-  GQ: 'Equatorial Guinea',
-  GR: 'Greece',
-  GT: 'Guatemala',
-  GU: 'Guam',
-  GW: 'Guinea-Bissau',
-  GY: 'Guyana',
-  HK: 'Hong Kong',
-  HN: 'Honduras',
-  HR: 'Croatia (Hrvatska)',
-  HT: 'Haiti',
-  HU: 'Hungary',
-  ID: 'Indonesia',
-  IE: 'Ireland',
-  IL: 'Israel',
-  IM: 'Isle of Man',
-  IN: 'India',
-  IQ: 'Iraq',
-  IR: 'Iran',
-  IS: 'Iceland',
-  IT: 'Italy',
-  JE: 'Jersey',
-  JM: 'Jamaica',
-  JO: 'Jordan',
-  JP: 'Japan',
-  KE: 'Kenya',
-  KG: 'Kyrgyzstan',
-  KH: 'Cambodia',
-  KI: 'Kiribati',
-  KM: 'Comoros',
-  KN: 'Saint Kitts and Nevis',
-  KP: 'North Korea',
-  KR: 'South Korea',
-  KW: 'Kuwait',
-  KY: 'Cayman Islands',
-  KZ: 'Kazakhstan',
-  LA: 'Lao',
-  LB: 'Lebanon',
-  LC: 'Saint Lucia',
-  LI: 'Liechtenstein',
-  LK: 'Sri Lanka',
-  LR: 'Liberia',
-  LS: 'Lesotho',
-  LT: 'Lithuania',
-  LU: 'Luxembourg',
-  LV: 'Latvia',
-  LY: 'Libyan Arab Jamahiriya',
-  MA: 'Morocco',
-  MC: 'Monaco',
-  MD: 'Moldova, Republic of',
-  ME: 'Montenegro',
-  MG: 'Madagascar',
-  MH: 'Marshall Islands',
-  MK: 'Macedonia',
-  ML: 'Mali',
-  MM: 'Myanmar',
-  MN: 'Mongolia',
-  MP: 'Northern Mariana Islands',
-  MQ: 'Martinique',
-  MR: 'Mauritania',
-  MS: 'Montserrat',
-  MT: 'Malta',
-  MU: 'Mauritius',
-  MV: 'Maldives',
-  MW: 'Malawi',
-  MX: 'Mexico',
-  MY: 'Malaysia',
-  MZ: 'Mozambique',
-  NA: 'Namibia',
-  NC: 'New Caledonia',
-  NE: 'Niger',
-  NG: 'Nigeria',
-  NI: 'Nicaragua',
-  NL: 'Netherlands',
-  NO: 'Norway',
-  NP: 'Nepal',
-  NR: 'Nauru',
-  NZ: 'New Zealand',
-  OM: 'Oman',
-  PA: 'Panama',
-  PE: 'Peru',
-  PF: 'French Polynesia',
-  PG: 'Papua New Guinea',
-  PH: 'Philippines',
-  PK: 'Pakistan',
-  PL: 'Poland',
-  PM: 'St. Pierre and Miquelon',
-  PR: 'Puerto Rico',
-  PS: 'Palestine',
-  PT: 'Portugal',
-  PW: 'Palau',
-  PY: 'Paraguay',
-  QA: 'Qatar',
-  RE: 'Reunion',
-  RO: 'Romania',
-  RS: 'Serbia',
-  RU: 'Russian Federation',
-  RW: 'Rwanda',
-  SA: 'Saudi Arabia',
-  SB: 'Solomon Islands',
-  SC: 'Seychelles',
-  SD: 'Sudan',
-  SE: 'Sweden',
-  SG: 'Singapore',
-  SH: 'St. Helena',
-  SI: 'Slovenia',
-  SJ: 'Svalbard and Jan Mayen Islands',
-  SK: 'Slovakia',
-  SL: 'Sierra Leone',
-  SM: 'San Marino',
-  SN: 'Senegal',
-  SO: 'Somalia',
-  SR: 'Suriname',
-  ST: 'Sao Tome and Principe',
-  SV: 'El Salvador',
-  SY: 'Syrian Arab Republic',
-  SZ: 'Swaziland',
-  TD: 'Chad',
-  TF: 'French Southern Territories',
-  TG: 'Togo',
-  TH: 'Thailand',
-  TJ: 'Tajikistan',
-  TK: 'Tokelau',
-  TM: 'Turkmenistan',
-  TN: 'Tunisia',
-  TO: 'Tonga',
-  TR: 'Turkey',
-  TT: 'Trinidad and Tobago',
-  TV: 'Tuvalu',
-  TW: 'Taiwan',
-  TZ: 'Tanzania',
-  UA: 'Ukraine',
-  UG: 'Uganda',
-  UM: 'United States minor outlying islands',
-  US: 'United States',
-  UY: 'Uruguay',
-  UZ: 'Uzbekistan',
-  VC: 'Saint Vincent and the Grenadines',
-  VE: 'Venezuela',
-  VI: 'Virgin Islands (U.S.)',
-  VN: 'Vietnam',
-  VU: 'Vanuatu',
-  WF: 'Wallis and Futuna Islands',
-  WS: 'Samoa',
-  XK: 'Kosovo',
-  YE: 'Yemen',
-  ZA: 'South Africa',
-  ZM: 'Zambia',
-  ZW: 'Zimbabwe',
-};
